External rules to remove duplicate code
