+++ static/js/index.ts
@@ -1,6 +1,7 @@
 export * from './modal';
 export * from './app';
 export * from './auth';
+import './syntaxRules';
 import './syntaxModesRules-en';
 import './syntaxModesRules-nl';
 import './tabs';
